requirements.txt
# =============================================================================
# AI Employee Gold Tier - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core Dependencies (Bronze Tier)
# -----------------------------------------------------------------------------
watchdog>=3.0.0

# -----------------------------------------------------------------------------
# Google API (Gmail Watcher & Sender - Silver Tier)
# -----------------------------------------------------------------------------
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# -----------------------------------------------------------------------------
# Playwright (Browser Automation - Silver/Gold Tier)
# -----------------------------------------------------------------------------
playwright>=1.40.0

# After installing, run: playwright install chromium

# -----------------------------------------------------------------------------
# Twitter API (Twitter Watcher & Poster - Gold Tier)
# -----------------------------------------------------------------------------
tweepy>=4.14.0

# -----------------------------------------------------------------------------
# Facebook/Meta API (Facebook Watcher & Poster - Gold Tier)
# -----------------------------------------------------------------------------
facebook-business>=18.0.0
requests>=2.31.0

# -----------------------------------------------------------------------------
# Scheduling (Silver Tier)
# -----------------------------------------------------------------------------
schedule>=1.2.0

# -----------------------------------------------------------------------------
# Environment Variables
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
requests>=2.31.0
retry>=0.9.2

# -----------------------------------------------------------------------------
# MCP Server (Odoo Integration - Gold Tier)
# -----------------------------------------------------------------------------
# mcp>=0.1.0  # Uncomment when MCP package is available

# -----------------------------------------------------------------------------
# Development Dependencies (Optional)
# -----------------------------------------------------------------------------
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
